# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kaka Portfolio API
  version: "v12.0.0"
  description: API for the Kaka portfolio management system
servers:
  - url: http://114.132.238.12
  - url: http://127.0.0.1:5003
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                device_id:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/register:
    post:
      operationId: register
      summary: Register with invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, invite_code]
              properties:
                username:
                  type: string
                password:
                  type: string
                invite_code:
                  type: string
                device_id:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/invite/validate:
    post:
      operationId: validateInviteCode
      summary: Validate invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/password/change:
    post:
      operationId: changePassword
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [old_password, new_password]
              properties:
                old_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/refresh:
    post:
      operationId: refreshSession
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                device_id:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/logout:
    post:
      operationId: logout
      summary: Logout
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/profile:
    post:
      operationId: updateProfile
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname:
                  type: string
                avatar:
                  type: string
      responses:
        "200":
          description: OK
  /api/auth/bootstrap_credentials:
    post:
      operationId: bootstrapCredentials
      summary: Bootstrap username/password for legacy users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
...
  /api/auth/me:
    get:
      operationId: getCurrentUser
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
...